feat(angular-platform): kangaroo platform module
A variety of features are "platform" specific, as in required for the common kangaroo contract. This
extracts them into their own module.
