+++ @kangaroo/angular-platform/package.json
@@ -1,13 +1,19 @@
 {
-  "name": "@kangaroo/angular-authn",
+  "$schema": "../../node_modules/ng-packagr/package.schema.json",
+  "name": "@kangaroo/angular-platform",
   "version": "1.0.0",
-  "description": "OAuth2 API Injectees for Angular2+",
+  "description": "API platform contracts required by the Kanagaroo ecosystem.",
   "license": "Apache-2.0",
   "author": {
     "name": "Michael Krotscheck",
     "email": "krotscheck@gmail.com",
     "url": "https://krotscheck.net/"
   },
+  "typings": "lib/kangaroo-angular-platform.d.ts",
+  "main": "lib/bundles/kangaroo-angular-platform.umd.js",
+  "module": "lib/esm5/kangaroo-angular-platform.js",
+  "es2015": "lib/esm2015/kangaroo-angular-platform.js",
+  "metadata": "lib/kangaroo-angular-platform.metadata.json",
   "scripts": {
     "test": "ng test --code-coverage",
     "lint": "ng lint",
@@ -23,10 +29,7 @@
         "es2017"
       ],
       "comments": "none",
-      "licensePath": "LICENSE",
-      "umdModuleIds": {
-        "@kangaroo/angular-platform": "@kangaroo/angular-platform"
-      }
+      "licensePath": "LICENSE"
     }
   },
   "devDependencies": {
@@ -35,7 +38,7 @@
     "@angular/compiler-cli": "~5.2.4",
     "@types/jasmine": "~2.8.6",
     "@types/node": "~8.9.0",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "~4.1.0",
     "jasmine-core": "~2.99.1",
     "karma": "~2.0.0",
     "karma-browserstack-launcher": "~1.3.0",
@@ -52,12 +55,9 @@
     "typescript": "~2.7.2"
   },
   "dependencies": {
-    "@angular/common": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@kangaroo/angular-platform": "1.0.0",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "@angular/common": "~5.2.4",
+    "@angular/core": "~5.2.4",
+    "rxjs": "~5.5.6",
+    "zone.js": "~0.8.20"
   }
 }
