+++ @kangaroo/angular-platform/src/api/abstract-resource.service.spec.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 Michael Krotscheck
+ * Copyright (c) 2018 Michael Krotscheck
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy
@@ -13,20 +13,28 @@
  *
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-import { Inject, Optional } from '@angular/core';
-import { CommonModel, SortOrder } from './model';
+import { Inject, InjectionToken, Optional } from '@angular/core';
 import { async, TestBed } from '@angular/core/testing';
-import { AbstractSubresourceService } from './abstract-subresource.service';
-import { ADMIN_API_ROOT, AdminApiRootProvider } from './admin-api-root';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
+import { CommonModel } from './common.model';
+import { SortOrder } from './sort-order.enum';
+import { AbstractResourceService } from './abstract-resource.service';
+import { ObservableInput } from 'rxjs/Observable';
 
 /**
- * Unit tests for the AbstractSubresourceService.
+ * Test API root.
  */
-describe('AbstractSubresourceService', () => {
+export const API_ROOT: InjectionToken<ObservableInput<string>> =
+  new InjectionToken<ObservableInput<string>>('API_ROOT');
+
+/**
+ * Unit tests for the AbstractResourceService.
+ */
+describe('AbstractResourceService', () => {
 
   /**
    * Private entity class.
@@ -36,38 +44,22 @@ describe('AbstractSubresourceService', () => {
   }
 
   /**
-   * Private parent entity class.
-   */
-  interface ParentEntity extends CommonModel {
-    name: string;
-  }
-
-  /**
    * Test class, exposing request.
    */
-  class TestEntityService extends AbstractSubresourceService<ParentEntity, TestEntity> {
-
-    constructor(http: HttpClient,
-                @Optional() @Inject(ADMIN_API_ROOT) apiRoot: AdminApiRootProvider) {
-      super('parent', 'test', http, apiRoot);
+  class TestEntityService extends AbstractResourceService<TestEntity> {
+    constructor(http: HttpClient, @Optional() @Inject(API_ROOT) apiRoot: ObservableInput<string>) {
+      super('test', http, apiRoot);
     }
   }
 
+  let service: TestEntityService;
+  let controller: HttpTestingController;
   const validEntity: TestEntity = {
     id: 'test_id',
     createdDate: 10000,
     modifiedDate: 20000,
     name: 'test_name'
   };
-  const parentEntity: TestEntity = {
-    id: 'parent_id',
-    createdDate: 10000,
-    modifiedDate: 20000,
-    name: 'test_name'
-  };
-
-  let service: TestEntityService;
-  let controller: HttpTestingController;
 
   describe('with an API root', () => {
     beforeEach(() => {
@@ -75,7 +67,7 @@ describe('AbstractSubresourceService', () => {
         providers: [
           TestEntityService,
           {
-            provide: ADMIN_API_ROOT,
+            provide: API_ROOT,
             useValue: Promise.resolve('http://example.com/v1')
           }
         ],
@@ -95,21 +87,11 @@ describe('AbstractSubresourceService', () => {
 
     describe('read()', () => {
       it('should issue a GET request', async(() => {
-        service.read(parentEntity, 'test_id').subscribe();
+        service.read('test_id').subscribe();
 
         controller.match((request) => {
           expect(request.method).toEqual('GET');
-          expect(request.url).toEqual('http://example.com/v1/parent/parent_id/test/test_id');
-          return true;
-        });
-      }));
-
-      it('should gracefully handle no parent entity', async(() => {
-        service.read(null, 'test_id').subscribe();
-
-        controller.match((request) => {
-          expect(request.method).toEqual('GET');
-          expect(request.url).toEqual('http://example.com/v1/parent//test/test_id');
+          expect(request.url).toEqual('http://example.com/v1/test/test_id');
           return true;
         });
       }));
@@ -117,7 +99,6 @@ describe('AbstractSubresourceService', () => {
   });
 
   describe('with no API root', () => {
-
     beforeEach(() => {
       TestBed.configureTestingModule({
         providers: [ TestEntityService ],
@@ -136,15 +117,14 @@ describe('AbstractSubresourceService', () => {
 
       it('should map filters to the query', async(() => {
         service
-          .browse(parentEntity,
-            {
-              'foo': 'bar',
-              'lol': 'cat'
-            })
+          .browse({
+            'foo': 'bar',
+            'lol': 'cat'
+          })
           .subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/parent/parent_id/test/');
+          expect(request.url).toEqual('/test/');
           expect(request.params.get('foo')).toEqual('bar');
           expect(request.params.get('lol')).toEqual('cat');
           return true;
@@ -152,40 +132,40 @@ describe('AbstractSubresourceService', () => {
       }));
 
       it('should map sort', async(() => {
-        service.browse(parentEntity, null, 'name').subscribe();
+        service.browse(null, 'name').subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/parent/parent_id/test/');
+          expect(request.url).toEqual('/test/');
           expect(request.params.get('sort')).toEqual('name');
           return true;
         });
       }));
 
       it('should map order', async(() => {
-        service.browse(parentEntity, null, null, SortOrder.Ascending).subscribe();
+        service.browse(null, null, SortOrder.Ascending).subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/parent/parent_id/test/');
+          expect(request.url).toEqual('/test/');
           expect(request.params.get('order')).toEqual('ASC');
           return true;
         });
       }));
 
       it('should map offset', async(() => {
-        service.browse(parentEntity, null, null, null, 10).subscribe();
+        service.browse(null, null, null, 10).subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/parent/parent_id/test/');
+          expect(request.url).toEqual('/test/');
           expect(request.params.get('offset')).toEqual('10');
           return true;
         });
       }));
 
       it('should map limit', async(() => {
-        service.browse(parentEntity, null, null, null, null, 10).subscribe();
+        service.browse(null, null, null, null, 10).subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/parent/parent_id/test/');
+          expect(request.url).toEqual('/test/');
           expect(request.params.get('limit')).toEqual('10');
           return true;
         });
@@ -195,10 +175,10 @@ describe('AbstractSubresourceService', () => {
     describe('search()', () => {
 
       it('should pass the query', async(() => {
-        service.search(parentEntity, 'query').subscribe();
+        service.search('query').subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/parent/parent_id/test/search');
+          expect(request.url).toEqual('/test/search');
           expect(request.params.get('q')).toEqual('query');
           return true;
         });
@@ -206,14 +186,14 @@ describe('AbstractSubresourceService', () => {
 
       it('should map filters to the query', async(() => {
         service
-          .search(parentEntity, 'query', {
+          .search('query', {
             'foo': 'bar',
             'lol': 'cat'
           })
           .subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/parent/parent_id/test/search');
+          expect(request.url).toEqual('/test/search');
           expect(request.params.get('q')).toEqual('query');
           expect(request.params.get('foo')).toEqual('bar');
           expect(request.params.get('lol')).toEqual('cat');
@@ -222,10 +202,10 @@ describe('AbstractSubresourceService', () => {
       }));
 
       it('should map offset', async(() => {
-        service.search(parentEntity, 'query', null, 10).subscribe();
+        service.search('query', null, 10).subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/parent/parent_id/test/search');
+          expect(request.url).toEqual('/test/search');
           expect(request.params.get('q')).toEqual('query');
           expect(request.params.get('offset')).toEqual('10');
           return true;
@@ -233,10 +213,10 @@ describe('AbstractSubresourceService', () => {
       }));
 
       it('should map limit', async(() => {
-        service.search(parentEntity, 'query', null, null, 10).subscribe();
+        service.search('query', null, null, 10).subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/parent/parent_id/test/search');
+          expect(request.url).toEqual('/test/search');
           expect(request.params.get('q')).toEqual('query');
           expect(request.params.get('limit')).toEqual('10');
           return true;
@@ -246,11 +226,11 @@ describe('AbstractSubresourceService', () => {
 
     describe('create()', () => {
       it('should issue a POST request', async(() => {
-        service.create(parentEntity, validEntity).subscribe();
+        service.create(validEntity).subscribe();
 
         controller.match((request) => {
           expect(request.method).toEqual('POST');
-          expect(request.url).toEqual('/parent/parent_id/test/');
+          expect(request.url).toEqual('/test/');
           expect(request.body).toEqual(validEntity);
           return true;
         });
@@ -259,21 +239,11 @@ describe('AbstractSubresourceService', () => {
 
     describe('read()', () => {
       it('should issue a GET request', async(() => {
-        service.read(parentEntity, 'test_id').subscribe();
+        service.read('test_id').subscribe();
 
         controller.match((request) => {
           expect(request.method).toEqual('GET');
-          expect(request.url).toEqual('/parent/parent_id/test/test_id');
-          return true;
-        });
-      }));
-
-      it('should gracefully handle no parent entity', async(() => {
-        service.read(null, 'test_id').subscribe();
-
-        controller.match((request) => {
-          expect(request.method).toEqual('GET');
-          expect(request.url).toEqual('/parent//test/test_id');
+          expect(request.url).toEqual('/test/test_id');
           return true;
         });
       }));
@@ -281,11 +251,11 @@ describe('AbstractSubresourceService', () => {
 
     describe('update()', () => {
       it('should issue a PUT request', async(() => {
-        service.update(parentEntity, validEntity).subscribe();
+        service.update(validEntity).subscribe();
 
         controller.match((request) => {
           expect(request.method).toEqual('PUT');
-          expect(request.url).toEqual('/parent/parent_id/test/test_id');
+          expect(request.url).toEqual('/test/test_id');
           expect(request.body).toEqual(validEntity);
           return true;
         });
@@ -294,11 +264,11 @@ describe('AbstractSubresourceService', () => {
 
     describe('destroy()', () => {
       it('should issue a DELETE request', async(() => {
-        service.destroy(parentEntity, validEntity).subscribe();
+        service.destroy(validEntity).subscribe();
 
         controller.match((request) => {
           expect(request.method).toEqual('DELETE');
-          expect(request.url).toEqual('/parent/parent_id/test/test_id');
+          expect(request.url).toEqual('/test/test_id');
           return true;
         });
       }));
