+++ @kangaroo/angular-platform/src/api/abstract-subresource.service.spec.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 Michael Krotscheck
+ * Copyright (c) 2018 Michael Krotscheck
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy
@@ -13,20 +13,28 @@
  *
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-import { Inject, Optional } from '@angular/core';
-import { AbstractResourceService } from './abstract-resource.service';
-import { CommonModel, SortOrder } from './model';
+import { Inject, InjectionToken, Optional } from '@angular/core';
 import { async, TestBed } from '@angular/core/testing';
-import { ADMIN_API_ROOT, AdminApiRootProvider } from './admin-api-root';
+import { AbstractSubresourceService } from './abstract-subresource.service';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
+import { CommonModel } from './common.model';
+import { SortOrder } from './sort-order.enum';
+import { ObservableInput } from 'rxjs/Observable';
 
 /**
- * Unit tests for the AbstractResourceService.
+ * Test API root.
  */
-describe('AbstractResourceService', () => {
+export const API_ROOT: InjectionToken<ObservableInput<string>> =
+  new InjectionToken<ObservableInput<string>>('API_ROOT');
+
+/**
+ * Unit tests for the AbstractSubresourceService.
+ */
+describe('AbstractSubresourceService', () => {
 
   /**
    * Private entity class.
@@ -36,22 +44,38 @@ describe('AbstractResourceService', () => {
   }
 
   /**
+   * Private parent entity class.
+   */
+  interface ParentEntity extends CommonModel {
+    name: string;
+  }
+
+  /**
    * Test class, exposing request.
    */
-  class TestEntityService extends AbstractResourceService<TestEntity> {
-    constructor(http: HttpClient, @Optional() @Inject(ADMIN_API_ROOT) apiRoot: AdminApiRootProvider) {
-      super('test', http, apiRoot);
+  class TestEntityService extends AbstractSubresourceService<ParentEntity, TestEntity> {
+
+    constructor(http: HttpClient,
+                @Optional() @Inject(API_ROOT) apiRoot: ObservableInput<string>) {
+      super('parent', 'test', http, apiRoot);
     }
   }
 
-  let service: TestEntityService;
-  let controller: HttpTestingController;
   const validEntity: TestEntity = {
     id: 'test_id',
     createdDate: 10000,
     modifiedDate: 20000,
     name: 'test_name'
   };
+  const parentEntity: TestEntity = {
+    id: 'parent_id',
+    createdDate: 10000,
+    modifiedDate: 20000,
+    name: 'test_name'
+  };
+
+  let service: TestEntityService;
+  let controller: HttpTestingController;
 
   describe('with an API root', () => {
     beforeEach(() => {
@@ -59,7 +83,7 @@ describe('AbstractResourceService', () => {
         providers: [
           TestEntityService,
           {
-            provide: ADMIN_API_ROOT,
+            provide: API_ROOT,
             useValue: Promise.resolve('http://example.com/v1')
           }
         ],
@@ -79,11 +103,21 @@ describe('AbstractResourceService', () => {
 
     describe('read()', () => {
       it('should issue a GET request', async(() => {
-        service.read('test_id').subscribe();
+        service.read(parentEntity, 'test_id').subscribe();
 
         controller.match((request) => {
           expect(request.method).toEqual('GET');
-          expect(request.url).toEqual('http://example.com/v1/test/test_id');
+          expect(request.url).toEqual('http://example.com/v1/parent/parent_id/test/test_id');
+          return true;
+        });
+      }));
+
+      it('should gracefully handle no parent entity', async(() => {
+        service.read(null, 'test_id').subscribe();
+
+        controller.match((request) => {
+          expect(request.method).toEqual('GET');
+          expect(request.url).toEqual('http://example.com/v1/parent//test/test_id');
           return true;
         });
       }));
@@ -91,6 +125,7 @@ describe('AbstractResourceService', () => {
   });
 
   describe('with no API root', () => {
+
     beforeEach(() => {
       TestBed.configureTestingModule({
         providers: [ TestEntityService ],
@@ -109,14 +144,15 @@ describe('AbstractResourceService', () => {
 
       it('should map filters to the query', async(() => {
         service
-          .browse({
-            'foo': 'bar',
-            'lol': 'cat'
-          })
+          .browse(parentEntity,
+            {
+              'foo': 'bar',
+              'lol': 'cat'
+            })
           .subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/test/');
+          expect(request.url).toEqual('/parent/parent_id/test/');
           expect(request.params.get('foo')).toEqual('bar');
           expect(request.params.get('lol')).toEqual('cat');
           return true;
@@ -124,40 +160,40 @@ describe('AbstractResourceService', () => {
       }));
 
       it('should map sort', async(() => {
-        service.browse(null, 'name').subscribe();
+        service.browse(parentEntity, null, 'name').subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/test/');
+          expect(request.url).toEqual('/parent/parent_id/test/');
           expect(request.params.get('sort')).toEqual('name');
           return true;
         });
       }));
 
       it('should map order', async(() => {
-        service.browse(null, null, SortOrder.Ascending).subscribe();
+        service.browse(parentEntity, null, null, SortOrder.Ascending).subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/test/');
+          expect(request.url).toEqual('/parent/parent_id/test/');
           expect(request.params.get('order')).toEqual('ASC');
           return true;
         });
       }));
 
       it('should map offset', async(() => {
-        service.browse(null, null, null, 10).subscribe();
+        service.browse(parentEntity, null, null, null, 10).subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/test/');
+          expect(request.url).toEqual('/parent/parent_id/test/');
           expect(request.params.get('offset')).toEqual('10');
           return true;
         });
       }));
 
       it('should map limit', async(() => {
-        service.browse(null, null, null, null, 10).subscribe();
+        service.browse(parentEntity, null, null, null, null, 10).subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/test/');
+          expect(request.url).toEqual('/parent/parent_id/test/');
           expect(request.params.get('limit')).toEqual('10');
           return true;
         });
@@ -167,10 +203,10 @@ describe('AbstractResourceService', () => {
     describe('search()', () => {
 
       it('should pass the query', async(() => {
-        service.search('query').subscribe();
+        service.search(parentEntity, 'query').subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/test/search');
+          expect(request.url).toEqual('/parent/parent_id/test/search');
           expect(request.params.get('q')).toEqual('query');
           return true;
         });
@@ -178,14 +214,14 @@ describe('AbstractResourceService', () => {
 
       it('should map filters to the query', async(() => {
         service
-          .search('query', {
+          .search(parentEntity, 'query', {
             'foo': 'bar',
             'lol': 'cat'
           })
           .subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/test/search');
+          expect(request.url).toEqual('/parent/parent_id/test/search');
           expect(request.params.get('q')).toEqual('query');
           expect(request.params.get('foo')).toEqual('bar');
           expect(request.params.get('lol')).toEqual('cat');
@@ -194,10 +230,10 @@ describe('AbstractResourceService', () => {
       }));
 
       it('should map offset', async(() => {
-        service.search('query', null, 10).subscribe();
+        service.search(parentEntity, 'query', null, 10).subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/test/search');
+          expect(request.url).toEqual('/parent/parent_id/test/search');
           expect(request.params.get('q')).toEqual('query');
           expect(request.params.get('offset')).toEqual('10');
           return true;
@@ -205,10 +241,10 @@ describe('AbstractResourceService', () => {
       }));
 
       it('should map limit', async(() => {
-        service.search('query', null, null, 10).subscribe();
+        service.search(parentEntity, 'query', null, null, 10).subscribe();
 
         controller.match((request) => {
-          expect(request.url).toEqual('/test/search');
+          expect(request.url).toEqual('/parent/parent_id/test/search');
           expect(request.params.get('q')).toEqual('query');
           expect(request.params.get('limit')).toEqual('10');
           return true;
@@ -218,11 +254,11 @@ describe('AbstractResourceService', () => {
 
     describe('create()', () => {
       it('should issue a POST request', async(() => {
-        service.create(validEntity).subscribe();
+        service.create(parentEntity, validEntity).subscribe();
 
         controller.match((request) => {
           expect(request.method).toEqual('POST');
-          expect(request.url).toEqual('/test/');
+          expect(request.url).toEqual('/parent/parent_id/test/');
           expect(request.body).toEqual(validEntity);
           return true;
         });
@@ -231,11 +267,21 @@ describe('AbstractResourceService', () => {
 
     describe('read()', () => {
       it('should issue a GET request', async(() => {
-        service.read('test_id').subscribe();
+        service.read(parentEntity, 'test_id').subscribe();
 
         controller.match((request) => {
           expect(request.method).toEqual('GET');
-          expect(request.url).toEqual('/test/test_id');
+          expect(request.url).toEqual('/parent/parent_id/test/test_id');
+          return true;
+        });
+      }));
+
+      it('should gracefully handle no parent entity', async(() => {
+        service.read(null, 'test_id').subscribe();
+
+        controller.match((request) => {
+          expect(request.method).toEqual('GET');
+          expect(request.url).toEqual('/parent//test/test_id');
           return true;
         });
       }));
@@ -243,11 +289,11 @@ describe('AbstractResourceService', () => {
 
     describe('update()', () => {
       it('should issue a PUT request', async(() => {
-        service.update(validEntity).subscribe();
+        service.update(parentEntity, validEntity).subscribe();
 
         controller.match((request) => {
           expect(request.method).toEqual('PUT');
-          expect(request.url).toEqual('/test/test_id');
+          expect(request.url).toEqual('/parent/parent_id/test/test_id');
           expect(request.body).toEqual(validEntity);
           return true;
         });
@@ -256,11 +302,11 @@ describe('AbstractResourceService', () => {
 
     describe('destroy()', () => {
       it('should issue a DELETE request', async(() => {
-        service.destroy(validEntity).subscribe();
+        service.destroy(parentEntity, validEntity).subscribe();
 
         controller.match((request) => {
           expect(request.method).toEqual('DELETE');
-          expect(request.url).toEqual('/test/test_id');
+          expect(request.url).toEqual('/parent/parent_id/test/test_id');
           return true;
         });
       }));
