+++ @kangaroo/angular-platform/src/api/common.model.ts
