+++ @kangaroo/angular-platform/src/api/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 Michael Krotscheck
+ * Copyright (c) 2018 Michael Krotscheck
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy
@@ -13,29 +13,11 @@
  *
  * See the License for the specific language governing permissions and
  * limitations under the License.
- */
-
-import { CommonModel } from '@kangaroo/angular-platform';
-
-/**
- * This interface describes an authenticator and its configuration.
  *
- * @author Michael Krotscheck
  */
-export interface Authenticator extends CommonModel {
 
-  /**
-   * Entity ID ID of the client to whom this authenticator is registered.
-   */
-  client: string;
-
-  /**
-   * The Authenticator type
-   */
-  type: string;
-
-  /**
-   * Configuration properties for this authenticator
-   */
-  configuration: Map<string, string>;
-}
+export { AbstractResourceService } from './abstract-resource.service';
+export { AbstractSubresourceService } from './abstract-subresource.service';
+export { CommonModel } from './common.model';
+export { ListResponse } from './list-response.model';
+export { SortOrder } from './sort-order.enum';
