+++ @kangaroo/angular-platform/src/api/list-response.model.ts
