+++ @kangaroo/angular-platform/src/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 Michael Krotscheck
+ * Copyright (c) 2018 Michael Krotscheck
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy
@@ -13,29 +13,25 @@
  *
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-import { CommonModel } from '@kangaroo/angular-platform';
+import { NgModule } from '@angular/core';
+import { CsrfHttpInterceptor } from './http/csrf.http-interceptor';
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
+
+export * from './api';
 
 /**
- * This interface describes an authenticator and its configuration.
+ * This module contains injectees and abstract classes that are consistent
+ * across the entire kangaroo ecosystem.
  *
  * @author Michael Krotscheck
  */
-export interface Authenticator extends CommonModel {
-
-  /**
-   * Entity ID ID of the client to whom this authenticator is registered.
-   */
-  client: string;
-
-  /**
-   * The Authenticator type
-   */
-  type: string;
-
-  /**
-   * Configuration properties for this authenticator
-   */
-  configuration: Map<string, string>;
+@NgModule({
+  providers: [
+    {provide: HTTP_INTERCEPTORS, multi: true, useClass: CsrfHttpInterceptor}
+  ]
+})
+export class KangarooPlatformModule {
 }
